requirements.txt
# Core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1

# v3 — Testing Software (real-time)
flask-socketio==5.3.6
eventlet==0.35.2
